# api/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Integracion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Mostrar usuarios",
                "operationId": "eaedf6aa84fde38a08bf5d87308169f7",
                "responses": {
                    "200": {
                        "description": "Se mostraran todos los usuarios."
                    },
                    "401": {
                        "description": "Necesitas loguear"
                    },
                    "404": {
                        "description": "Contenido no encontrado"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crear usuarios",
                "operationId": "e2308aae8643f8f4d79821a6399114f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "documentacion_1"
                                    },
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "documentacion_1@gmail.com"
                                    },
                                    "nombre": {
                                        "description": "Nombre del individuo",
                                        "type": "string",
                                        "format": "string",
                                        "example": "documentacion_1"
                                    },
                                    "direccion": {
                                        "description": "Direccion del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "Documentando una api"
                                    },
                                    "id_municipio": {
                                        "description": "Id del municipio",
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "password": {
                                        "description": "Password del usuario",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "telefono": {
                                        "description": "Telefono del usuario",
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "123456789"
                                    },
                                    "avatar": {
                                        "description": "Avatar del usuario",
                                        "type": "string",
                                        "format": "ejemplo.png",
                                        "example": "documentacion_1.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se ha creado el usuario."
                    },
                    "401": {
                        "description": "Necesitas loguear"
                    },
                    "404": {
                        "description": "Contenido no encontrado"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualizar usuarios",
                "operationId": "15f5757522724ebca9bac30479321b45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del elemento a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "readOnly": true
                        },
                        "example": 14
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "documentacion_cambio"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "password",
                                        "example": "password"
                                    },
                                    "nombre": {
                                        "description": "Nombre identificativo",
                                        "type": "string",
                                        "example": "Documentacion Api"
                                    },
                                    "email": {
                                        "description": "Nombre del usuario",
                                        "type": "email",
                                        "example": "documentacion_cambio@email.com"
                                    },
                                    "direccion": {
                                        "description": "Direccion del usuario",
                                        "type": "string",
                                        "example": "Documentacion Cambio Direccion"
                                    },
                                    "id_municipio": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "telefono": {
                                        "description": "Telefono del usuario",
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "avatar": {
                                        "description": "Avatar del usuario",
                                        "type": "string",
                                        "example": "documentacion.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se ha actualizado el usuario."
                    },
                    "404": {
                        "description": "No se ha encontrado el id del usuario o el usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Error de validacion del usuario."
                    },
                    "500": {
                        "description": "No se ha podido realizar el cambio en el servidor."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Borra usuarios",
                "operationId": "98b6e09e3d231384330198299feb017f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del elemento a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Se ha creado el usuario."
                    },
                    "403": {
                        "description": "Error de permisos."
                    },
                    "404": {
                        "description": "No se ha encontrado el elemento."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Necesitas loguear."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer Token)",
                "name": "Authorization",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}